refactor(identity): separate normalized policy from env adapter
